Move bigint benchmark to criterion
